Move helper method
This is now only used by the world index presenter so the method has been moved there and the helper file deleted
