--- app/helpers/world_location_helper.rb
@@ -1,8 +1,0 @@
-module WorldLocationHelper
-  def sort(locations)
-    locations
-      .map { |location| [ActiveSupport::Inflector.transliterate(location.name_without_prefix), location] }
-      .sort_by { |transliterated_name, _location| transliterated_name.downcase }
-      .map { |_transliterated_name, location| location }
-  end
-end
