+++ test/unit/app/presenters/publishing_api/world_index_presenter_test.rb
@@ -91,4 +91,64 @@ class PublishingApi::WorldIndexPresenterTest < ActiveSupport::TestCase
     assert_equal expected_links, presenter.links
     assert_valid_against_links_schema({ links: presenter.links }, "world_index")
   end
+
+  test "sorts correctly where special characters are used anywhere other than the first character" do
+    world_location_1 = create(:world_location, name: "Cow")
+    world_location_2 = create(:world_location, name: "Côt")
+    unsorted = [world_location_1, world_location_2]
+
+    expected_order = [
+      world_location_2,
+      world_location_1,
+    ]
+
+    presenter = PublishingApi::WorldIndexPresenter.new
+
+    assert_equal expected_order, presenter.sort(unsorted)
+  end
+
+  test "sorts correctly where special characters are used as the first character" do
+    world_location_1 = create(:world_location, name: "England")
+    world_location_2 = create(:world_location, name: "Éire")
+    unsorted = [world_location_1, world_location_2]
+
+    expected_order = [
+      world_location_2,
+      world_location_1,
+    ]
+
+    presenter = PublishingApi::WorldIndexPresenter.new
+
+    assert_equal expected_order, presenter.sort(unsorted)
+  end
+
+  test "ignores `The` prefix when sorting" do
+    world_location_1 = create(:world_location, name: "Hungary")
+    world_location_2 = create(:world_location, name: "The Gambia")
+    unsorted = [world_location_1, world_location_2]
+
+    expected_order = [
+      world_location_2,
+      world_location_1,
+    ]
+
+    presenter = PublishingApi::WorldIndexPresenter.new
+
+    assert_equal expected_order, presenter.sort(unsorted)
+  end
+
+  test "ignores case when sorting" do
+    world_location_1 = create(:world_location, name: "UK Mission to ASEAN")
+    world_location_2 = create(:world_location, name: "UK and the Commonwealth")
+    unsorted = [world_location_1, world_location_2]
+
+    expected_order = [
+      world_location_2,
+      world_location_1,
+    ]
+
+    presenter = PublishingApi::WorldIndexPresenter.new
+
+    assert_equal expected_order, presenter.sort(unsorted)
+  end
 end
